chore: move status to own file
