+++ lapce-app/src/lib.rs
@@ -28,6 +28,7 @@ pub mod rename;
 pub mod settings;
 pub mod snippet;
 pub mod source_control;
+pub mod status;
 pub mod terminal;
 pub mod text_area;
 pub mod text_input;
